+++ src/util/omit.ts
@@ -1,5 +1,5 @@
 export default function omit(source: any, ...keys: string[]): any {
-    return Object.keys(source).reduce((acc, key) => {
-        return keys.indexOf(key) === -1 ? { ...acc, [key]: source[key] } : acc
-    }, {})
+  return Object.keys(source).reduce((acc, key) => {
+    return keys.indexOf(key) === -1 ? { ...acc, [key]: source[key] } : acc
+  }, {})
 }