+++ src/util/setData.ts
@@ -1,8 +1,8 @@
-export default function setData({ id, data }: { id: string, data: any }) {
-    return (state: any) => ({
-        [id]: {
-            ...state[id],
-            ...data
-        }
-    })
+export default function setData({ id, data }: { id: string; data: any }) {
+  return (state: any) => ({
+    [id]: {
+      ...state[id],
+      ...data
+    }
+  })
 }